+++ src/qimcifa.cpp
@@ -54,327 +54,9 @@
 // See LICENSE.md in the project root or https://www.gnu.org/licenses/lgpl-3.0.en.html
 // for details.
 
-#include "config.h"
-
-#include <algorithm>
-#include <chrono>
-#include <cmath>
-#include <float.h>
-#include <fstream>
-#include <future>
-#include <iomanip>
-#include <iostream>
-#include <map>
-#include <mutex>
-#include <stdlib.h>
-#include <string>
-#include <time.h>
-
-#include <boost/dynamic_bitset.hpp>
-
-#if USE_GMP
-#include <boost/multiprecision/gmp.hpp>
-#elif USE_BOOST
-#include <boost/multiprecision/cpp_int.hpp>
-#else
-#include "big_integer.hpp"
-#endif
+#include "qimcifa.hpp"
 
 namespace Qimcifa {
-
-// Make this a multiple of 2, 3, 5, 7, 11, 13, and 17.
-// constexpr int BIGGEST_WHEEL = 510510;
-constexpr int BIGGEST_WHEEL = 1021020;
-
-#if USE_GMP
-typedef boost::multiprecision::mpz_int BigIntegerInput;
-#elif USE_BOOST
-typedef boost::multiprecision::number<boost::multiprecision::cpp_int_backend<4096, 4096,
-    boost::multiprecision::unsigned_magnitude, boost::multiprecision::unchecked, void>>
-    BigIntegerInput;
-#else
-typedef BigInteger BigIntegerInput;
-#endif
-
-#if IS_PARALLEL
-BigIntegerInput batchNumber;
-BigIntegerInput batchCount;
-std::mutex batchMutex;
-
-inline void finish() {
-    std::lock_guard<std::mutex> lock(batchMutex);
-    batchNumber = batchCount;
-}
-
-inline BigIntegerInput getNextBatch() {
-    std::lock_guard<std::mutex> lock(batchMutex);
-    BigIntegerInput result = batchNumber;
-    if (batchNumber < batchCount) {
-        ++batchNumber;
-    }
-
-    return result;
-}
-#endif
-
-// See https://stackoverflow.com/questions/101439/the-most-efficient-way-to-implement-an-integer-based-power-function-powint-int
-template <typename BigInteger> BigInteger ipow(BigInteger base, unsigned exp)
-{
-    BigInteger result = 1U;
-    for (;;)
-    {
-        if (exp & 1U) {
-            result *= base;
-        }
-        exp >>= 1U;
-        if (!exp) {
-            break;
-        }
-        base *= base;
-    }
-
-    return result;
-}
-
-template <typename BigInteger> inline BigInteger sqrt(const BigInteger& toTest)
-{
-    // Otherwise, find b = sqrt(b^2).
-    BigInteger start = 1U, end = toTest >> 1U, ans = 0U;
-    do {
-        const BigInteger mid = (start + end) >> 1U;
-
-        // If toTest is a perfect square
-        const BigInteger sqr = mid * mid;
-        if (sqr == toTest) {
-            ans = mid;
-            break;
-        }
-
-        if (sqr < toTest) {
-            // Since we need floor, we update answer when mid*mid is smaller than p, and move closer to sqrt(p).
-            start = mid + 1U;
-            ans = mid;
-        } else {
-            // If mid*mid is greater than p
-            end = mid - 1U;
-        }
-    } while (start <= end);
-
-    return ans;
-}
-
-template <typename BigInteger> inline uint64_t log2(BigInteger n) {
-#if USE_GMP || USE_BOOST
-    uint64_t pow = 0U;
-    BigInteger p = n >> 1U;
-    while (p) {
-        p >>= 1U;
-        ++pow;
-    }
-    return pow;
-#else
-    return bi_log2(n);
-#endif
-}
-
-template <typename BigInteger> inline bool isPowerOfTwo(const BigInteger& x)
-{
-    // Source: https://www.exploringbinary.com/ten-ways-to-check-if-an-integer-is-a-power-of-two-in-c/
-#if USE_GMP || USE_BOOST
-    return (x && !(x & (x - 1ULL)));
-#else
-    BigInteger y = x;
-    bi_decrement(&y, 1U);
-    bi_and_ip(&y, x);
-    return (bi_compare_0(x) != 0) && (bi_compare_0(y) == 0);
-#endif
-}
-
-template <typename BigInteger> inline BigInteger gcd(BigInteger n1, BigInteger n2)
-{
-    while (n2 != 0) {
-        const BigInteger t = n1;
-        n1 = n2;
-        n2 = t % n2;
-    }
-
-    return n1;
-}
-
-template <typename BigInteger>
-void printSuccess(const BigInteger& f1, const BigInteger& f2, const BigInteger& toFactor, const std::string& message,
-    const std::chrono::time_point<std::chrono::high_resolution_clock>& iterClock)
-{
-    std::cout << message << f1 << " * " << f2 << " = " << toFactor << std::endl;
-    auto tClock =
-        std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::high_resolution_clock::now() - iterClock);
-    // Report in seconds
-    std::cout << "(Time elapsed: " << (tClock.count() / 1000000.0) << " seconds)" << std::endl;
-    std::cout << "(Waiting to join other threads...)" << std::endl;
-}
-
-template <typename BigInteger> inline BigInteger backward(BigInteger n) {
-    return ((~((~n) | 1U)) / 3U) + 1U;
-}
-
-template <typename BigInteger> inline BigInteger forward(BigInteger p) {
-    // Make this NOT a multiple of 2 or 3.
-    return (p << 1U) + (~(~p | 1U)) - 1U;
-}
-
-template <typename BigInteger> bool isMultiple(const BigInteger& p, const std::vector<BigInteger>& knownPrimes) {
-    for (const BigInteger& prime : knownPrimes) {
-        if ((p % prime) == 0) {
-            return true;
-        }
-    }
-    return false;
-}
-
-template <typename BigInteger>
-boost::dynamic_bitset<size_t> wheel_inc(std::vector<BigInteger> primes, BigInteger limit) {
-    BigInteger radius = 1U;
-    for (const BigInteger& i : primes) {
-        radius *= i;
-    }
-    if (limit < radius) {
-        radius = limit;
-    }
-    const BigInteger prime = primes.back();
-    primes.pop_back();
-    std::vector<bool> o;
-    for (BigInteger i = 1U; i <= radius; ++i) {
-        if (!isMultiple(i, primes)) {
-            o.push_back((i % prime) == 0);
-        }
-    }
-
-    boost::dynamic_bitset<size_t> output(o.size());
-    for (size_t i = 0U; i < o.size(); ++i) {
-        output[i] = o[i];
-    }
-    output >>= 1U;
-
-    return output;
-}
-
-template <typename BigInteger>
-std::vector<boost::dynamic_bitset<size_t>> wheel_gen(const std::vector<BigInteger>& primes, BigInteger limit) {
-    std::vector<boost::dynamic_bitset<size_t>> output;
-    std::vector<BigInteger> wheelPrimes;
-    for (const BigInteger& p : primes) {
-        wheelPrimes.push_back(p);
-        output.push_back(wheel_inc(wheelPrimes, limit));
-    }
-    return output;
-}
-
-inline size_t GetWheelIncrement(std::vector<boost::dynamic_bitset<size_t>>& inc_seqs) {
-    size_t wheelIncrement = 0U;
-    bool is_wheel_multiple = false;
-    do {
-        for (size_t i = 0; i < inc_seqs.size(); ++i) {
-            boost::dynamic_bitset<size_t>& wheel = inc_seqs[i];
-            is_wheel_multiple = wheel.test(0U);
-            wheel >>= 1U;
-            if (is_wheel_multiple) {
-                wheel[wheel.size() - 1U] = true;
-                break;
-            }
-        }
-        wheelIncrement++;
-    } while (is_wheel_multiple);
-
-    return wheelIncrement;
-}
-
-template <typename BigInteger>
-inline bool checkCongruenceOfSquares(const BigInteger& toFactor, const BigInteger& toTest, const BigInteger& radius,
-    const std::chrono::time_point<std::chrono::high_resolution_clock>& iterClock)
-{
-    // The basic idea is "congruence of squares":
-    // a^2 = b^2 mod N
-    // If we're lucky enough that the above is true, for a^2 = toTest and (b^2 mod N) = remainder,
-    // then we can immediately find a factor.
-
-    // Consider a to be equal to "toTest."
-    const BigInteger bSqr = (toTest * toTest) % toFactor;
-    const BigInteger b = sqrt(bSqr);
-    if ((b * b) != bSqr) {
-        if (bSqr < radius) {
-            std::cout << toTest << std::endl;
-        }
-        return false;
-    }
-
-    BigInteger f1 = gcd(toTest + b, toFactor);
-    BigInteger f2 = gcd(toTest - b, toFactor);
-    BigInteger fmul = f1 * f2;
-    while ((fmul > 1U) && (fmul != toFactor) && ((toFactor % fmul) == 0)) {
-        fmul = f1;
-        f1 = f1 * f2;
-        f2 = toFactor / (fmul * f2);
-        fmul = f1 * f2;
-    }
-    if ((fmul == toFactor) && (f1 > 1U) && (f2 > 1U)) {
-        // Inform the other threads on this node that we've succeeded and are done:
-        printSuccess<BigInteger>(f1, f2, toFactor, "Congruence of squares: Found ", iterClock);
-        return true;
-    }
-
-    return false;
-}
-
-template <typename BigInteger>
-inline bool getSmoothNumbersIteration(const BigInteger& toFactor, const BigInteger& base, const BigInteger& radius,
-    const std::chrono::time_point<std::chrono::high_resolution_clock>& iterClock) {
-#if IS_RSA_SEMIPRIME
-    if ((toFactor % base) == 0U) {
-        printSuccess<BigInteger>(base, toFactor / base, toFactor, "Exact factor: Found ", iterClock);
-        return true;
-    }
-#else
-    BigInteger n = gcd(base, toFactor);
-    if (n != 1U) {
-        printSuccess<BigInteger>(n, toFactor / n, toFactor, "Has common factor: Found ", iterClock);
-        return true;
-    }
-#endif
-
-    return checkCongruenceOfSquares<BigInteger>(toFactor, base, radius, iterClock);
-}
-
-#if IS_PARALLEL
-template <typename BigInteger>
-bool getSmoothNumbers(const BigInteger& toFactor, std::vector<boost::dynamic_bitset<uint64_t>> inc_seqs, const BigInteger& radius,
-    const std::chrono::time_point<std::chrono::high_resolution_clock>& iterClock)
-{
-    for (BigInteger batchNum = (BigInteger)getNextBatch(); batchNum < batchCount; batchNum = (BigInteger)getNextBatch()) {
-        for (size_t batchItem = 0U; batchItem < BIGGEST_WHEEL;) {
-            batchItem += GetWheelIncrement(inc_seqs);
-            if (getSmoothNumbersIteration<BigInteger>(toFactor, forward(batchItem), radius, iterClock)) {
-                return true;
-            }
-        }
-    }
-
-    return false;
-}
-#else
-template <typename BigInteger>
-bool getSmoothNumbers(const BigInteger& toFactor, std::vector<boost::dynamic_bitset<uint64_t>> inc_seqs, const BigInteger& offset,
-    const BigInteger& range, const BigInteger& radius, const std::chrono::time_point<std::chrono::high_resolution_clock>& iterClock)
-{
-    for (BigInteger batchItem = offset; batchItem < range;) {
-        batchItem += GetWheelIncrement(inc_seqs);
-        if (getSmoothNumbersIteration<BigInteger>(toFactor, forward(batchItem), radius, iterClock)) {
-            return true;
-        }
-    }
-
-    return false;
-}
-#endif
 
 template <typename BigInteger>
 int mainBody(const BigInteger& toFactor)
@@ -442,7 +124,48 @@ int mainBody(const BigInteger& toFactor)
     //     std::cout << i << ": " << trialDivisionPrimes[i] << ", ";
     // }
 
-    const size_t tdLevel = 7U;
+#if IS_PARALLEL
+    const unsigned cpuCount = std::thread::hardware_concurrency();
+#else
+    const unsigned cpuCount = 1U;
+#endif
+
+    int64_t tdLevel = 7;
+    std::cout << "Wheel factorization level (minimum of " << MIN_RTD_LEVEL << ", max of 7, or -1 for calibration file): ";
+    std::cin >> tdLevel;
+    if ((tdLevel > -1) && (tdLevel < MIN_RTD_LEVEL)) {
+        tdLevel = MIN_RTD_LEVEL;
+    }
+    if (tdLevel > 7) {
+        tdLevel = 7;
+    }
+    if (tdLevel < 0) {
+        std::ifstream settingsFile ("qimcifa_calibration.ssv");
+        std::string header;
+        std::getline(settingsFile, header);
+        size_t bestLevel = MIN_RTD_LEVEL;
+        double bestCost = DBL_MAX;
+        while (settingsFile.peek() != EOF)
+        {
+            size_t level;
+            BigInteger cardinality;
+            double batchTime, cost;
+            settingsFile >> level;
+            settingsFile >> cardinality;
+            settingsFile >> batchTime;
+            settingsFile >> cost;
+
+            if ((level >= MIN_RTD_LEVEL) && (cost < bestCost)) {
+                tdLevel = level;
+                bestCost = cost;
+            }
+        }
+        settingsFile.close();
+
+        std::cout << "Calibrated reverse trial division level: " << bestLevel << std::endl;
+        std::cout << "Estimated average time to exit: " << (bestCost / (2 * cpuCount)) << " seconds" << std::endl;
+    }
+
     auto iterClock = std::chrono::high_resolution_clock::now();
     const BigInteger fullMaxBase = sqrt<BigInteger>(toFactor);
     if (fullMaxBase * fullMaxBase == toFactor) {
@@ -450,7 +173,7 @@ int mainBody(const BigInteger& toFactor)
         return 0;
     }
 
-    for (size_t primeIndex = 0; primeIndex < tdLevel; ++primeIndex) {
+    for (int64_t primeIndex = 0; primeIndex < tdLevel; ++primeIndex) {
         const unsigned currentPrime = trialDivisionPrimes[primeIndex];
         if ((toFactor % currentPrime) == 0) {
             std::cout << "Factors: " << currentPrime << " * " << (toFactor / currentPrime) << " = " << toFactor
@@ -466,14 +189,21 @@ int mainBody(const BigInteger& toFactor)
     std::vector<boost::dynamic_bitset<uint64_t>> inc_seqs = wheel_gen(std::vector<BigInteger>(trialDivisionPrimes.begin(), trialDivisionPrimes.begin() + tdLevel), toFactor);
     inc_seqs.erase(inc_seqs.begin(), inc_seqs.begin() + 2U);
 
+#if BIG_INTEGER_BITS > 64 && !USE_BOOST && !USE_GMP
+    const double exp = log2(bi_to_double(toFactor));
+#elif BIG_INTEGER_BITS > 64 && USE_BOOST || USE_GMP
+    const double exp = log2(toFactor.convert_to<double>());
+#else
+    const double exp = log2(toFactor);
+#endif
+
     BigInteger radius = 1U;
-    for (size_t i = 0U; i < tdLevel; ++i) {
+    for (int64_t i = 0U; i < tdLevel; ++i) {
         radius *= trialDivisionPrimes[i];
     }
-    radius *= radius;
+    radius = (BigInteger)pow((uint64_t)radius, exp / 32.0);
 
 #if IS_PARALLEL
-    const unsigned cpuCount = std::thread::hardware_concurrency();
     const BigInteger threadRange = ((((fullRange + cpuCount - 1U) / cpuCount) + BIGGEST_WHEEL - 1U) / BIGGEST_WHEEL) * BIGGEST_WHEEL;
     batchNumber = 0;
     batchCount = (cpuCount * threadRange) / BIGGEST_WHEEL;
@@ -489,7 +219,7 @@ int mainBody(const BigInteger& toFactor)
     }
 #else
     std::vector<BigInteger> smoothNumbers;
-    getSmoothNumbers(toFactor, inc_seqs, offset, fullRange, radius, iterClock);
+    getSmoothNumbers(toFactor, inc_seqs, offset, radius, fullRange, iterClock);
 #endif
 
     return 0;
